refactor: Move remaining error types to telemetry-utils
