+++ packages/loader/container-utils/src/index.ts
@@ -4,12 +4,7 @@
  */
 
 export { DeltaManagerProxyBase } from "./deltaManagerProxyBase";
-export {
-	ClientSessionExpiredError,
-	DataProcessingError,
-	extractSafePropertiesFromMessage,
-	ThrottlingWarning,
-} from "./error";
+export { ClientSessionExpiredError, ThrottlingWarning } from "./error";
 
 // Deprecated exports for backwards compatibility
 export {
@@ -17,6 +12,14 @@ export {
 	 * @deprecated Import from `@fluidframework/telemetry-utils` instead.
 	 */
 	DataCorruptionError,
+	/**
+	 * @deprecated Import from `@fluidframework/telemetry-utils` instead.
+	 */
+	DataProcessingError,
+	/**
+	 * @deprecated Import from `@fluidframework/telemetry-utils` instead.
+	 */
+	extractSafePropertiesFromMessage,
 	/**
 	 * @deprecated Import from `@fluidframework/telemetry-utils` instead.
 	 */
